refactor(keybinds): Let increase and decrease font size working with shortcuts with new way
