+++ ui/src/components/shortcuts/mod.rs
@@ -1,2 +1,1 @@
-pub mod change_font_size_shortcut;
 pub mod paste_file_shortcut;
