+++ ui/src/lib.rs
@@ -48,7 +48,6 @@ use std::sync::Arc;
 
 use crate::auth_guard::AuthGuard;
 use crate::components::debug_logger::DebugLogger;
-use crate::components::shortcuts::change_font_size_shortcut::ChangeFontSizeShortCut;
 use crate::components::toast::Toast;
 use crate::components::topbar::release_info::Release_Info;
 use crate::layouts::community::CommunityLayout;
@@ -231,6 +230,8 @@ fn app_layout(cx: Scope) -> Element {
     use_app_coroutines(cx)?;
     use_router_notification_listener(cx)?;
 
+    let state = use_shared_state::<State>(cx)?;
+
     render! {
         AppStyle {}
         div { id: "app-wrap",
@@ -240,8 +241,8 @@ fn app_layout(cx: Scope) -> Element {
                     match shortcut {
                         GlobalShortcut::ToggleMute => utils::keyboard::shortcut_handlers::audio::toggle_mute(),
                         GlobalShortcut::ToggleDeafen => utils::keyboard::shortcut_handlers::audio::toggle_deafen(),
-                        GlobalShortcut::IncreaseFontSize => utils::keyboard::shortcut_handlers::font::increase_size(),
-                        GlobalShortcut::DecreaseFontSize => utils::keyboard::shortcut_handlers::font::decrease_size(),
+                        GlobalShortcut::IncreaseFontSize => utils::keyboard::shortcut_handlers::font::increase_size(state.clone()),
+                        GlobalShortcut::DecreaseFontSize => utils::keyboard::shortcut_handlers::font::decrease_size(state.clone()),
                     }
                     log::debug!("shortcut called {:?}", shortcut);
                 }
@@ -1174,37 +1175,32 @@ fn AppNav<'a>(
     };
     let _routes = vec![chat_route, files_route, friends_route, settings_route];
 
-    render!(
-        if state.read().ui.metadata.focused {
-            rsx!(ChangeFontSizeShortCut {})
-        }
-        kit::components::nav::Nav {
-            routes: _routes,
-            active: match active {
-                UplinkRoute::ChatLayout {} => "/chat",
-                UplinkRoute::SettingsLayout {} => "/settings",
-                UplinkRoute::FriendsLayout {} => "/friends",
-                UplinkRoute::FilesLayout {} => "/files",
-                _ => "",
-            },
-            onnavigate: move |r| {
-                if let Some(f) = onnavigate {
-                    f.call(());
-                }
+    render!(kit::components::nav::Nav {
+        routes: _routes,
+        active: match active {
+            UplinkRoute::ChatLayout {} => "/chat",
+            UplinkRoute::SettingsLayout {} => "/settings",
+            UplinkRoute::FriendsLayout {} => "/friends",
+            UplinkRoute::FilesLayout {} => "/files",
+            _ => "",
+        },
+        onnavigate: move |r| {
+            if let Some(f) = onnavigate {
+                f.call(());
+            }
 
-                let new_layout = match r {
-                    "/chat" => UplinkRoute::ChatLayout {},
-                    "/settings" => UplinkRoute::SettingsLayout {},
-                    "/friends" => UplinkRoute::FriendsLayout {},
-                    "/files" => UplinkRoute::FilesLayout {},
-                    _ => UplinkRoute::ChatLayout {},
-                };
+            let new_layout = match r {
+                "/chat" => UplinkRoute::ChatLayout {},
+                "/settings" => UplinkRoute::SettingsLayout {},
+                "/friends" => UplinkRoute::FriendsLayout {},
+                "/files" => UplinkRoute::FilesLayout {},
+                _ => UplinkRoute::ChatLayout {},
+            };
 
-                navigator.replace(new_layout);
-            },
-            tooltip_direction: tooltip_direction.unwrap_or(ArrowPosition::Bottom),
-        }
-    )
+            navigator.replace(new_layout);
+        },
+        tooltip_direction: tooltip_direction.unwrap_or(ArrowPosition::Bottom),
+    })
 }
 
 struct LogDropper {}
